+++ lib/glimmer/libui/control_proxy/column.rb
@@ -19,26 +19,37 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-require 'glimmer/libui/control_proxy'
-require 'glimmer/libui/column'
-require 'glimmer/libui/editable_column'
-
 module Glimmer
   module LibUI
     class ControlProxy
-      # Proxy for LibUI checkbox column objects
-      #
-      # Follows the Proxy Design Pattern
-      class CheckboxColumnProxy < ControlProxy
-        include Column
-        include EditableColumn
+      # Common logic for all column proxy objects
+      module Column
+        def initialize(keyword, parent, args, &block)
+          @keyword = keyword
+          @parent_proxy = parent
+          @args = args
+          @block = block
+          @enabled = true
+          post_add_content if @block.nil?
+        end
         
-        private
+        def name
+          @args.first
+        end
         
-        def build_control
-          @parent_proxy.append_checkbox_column(name, column_index, editable_value)
+        # column index used in table append column API call
+        # expanded to ensure DualColumn index accounts for two columns acting as one
+        def column_index
+          @column_index ||= @parent_proxy.send(:next_column_index)
+        end
+        
+        # actual index within table columns (disregarding nil fillings that account for DualColumn instances)
+        def index
+          @parent_proxy.columns.compact.index(self)
         end
       end
     end
   end
 end
+
+Dir[File.expand_path('./column/*.rb', __dir__)].each {|f| require f}