+++ lib/glimmer/libui/control_proxy/enableable_column.rb
@@ -21,23 +21,25 @@
 
 module Glimmer
   module LibUI
-    module EditableColumn
-      def editable(value = nil)
-        if value.nil?
-          @editable = false if @editable.nil?
-          @editable
-        else
-          @editable = !!value
+    class ControlProxy
+      module EnableableColumn
+        def enabled(value = nil)
+          if value.nil?
+            @enabled = true if @enabled.nil?
+            @enabled
+          else
+            @enabled = !!value
+          end
         end
-      end
-      alias editable= editable
-      alias set_editable editable
-      alias editable? editable
-
-      private
-      
-      def editable_value
-        (@parent_proxy.editable? || editable?) ? -2 : -1
+        alias enabled= enabled
+        alias set_enabled enabled
+        alias enabled? enabled
+  
+        private
+        
+        def enabled_value
+          enabled? ? -2 : -1
+        end
       end
     end
   end